pyproject.toml
[project]
name = "ml-demo-2026"
version = "0.1.0"
description = ""
authors = [
    {name = "Adrian Vega",email = "adrian.vega.del@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.serve.dependencies]
fastapi = "^0.133.1"
uvicorn = "^0.41.0"
pydantic = "^2.12.5"
google-cloud-bigquery = "^3.40.1"


[tool.poetry.group.test.dependencies]
requests = "^2.32.5"


[tool.poetry.group.local.dependencies]
python-dotenv = "^1.2.1"

